Remaining tseries tests reorg
