+++ pandas/tests/tseries/data/cday-0.14.1.pickle
