+++ pandas/tests/tseries/data/dateoffset_0_15_2.pickle
