+++ pandas/tests/tseries/test_bin_groupby.py
