+++ pandas/tests/tseries/test_frequencies.py
