+++ pandas/tests/tseries/test_resample.py
