+++ pandas/tests/tseries/test_timezones.py
