+++ setup.py
@@ -648,13 +648,13 @@ setup(name=DISTNAME,
                 'pandas.tests.series',
                 'pandas.tests.formats',
                 'pandas.tests.scalar',
+                'pandas.tests.tseries',
                 'pandas.tests.types',
                 'pandas.tests.test_msgpack',
                 'pandas.tests.plotting',
                 'pandas.tools',
                 'pandas.tools.tests',
                 'pandas.tseries',
-                'pandas.tseries.tests',
                 'pandas.types',
                 'pandas.io.tests',
                 'pandas.io.tests.json',
